style: изменить название пакета
