+++ ewm-service/src/main/java/ru/practicum/dto/compilation/NewCompilationDto.java
@@ -1,4 +1,4 @@
-package ru.practicum.api.compilation;
+package ru.practicum.dto.compilation;
 
 import lombok.Data;
