+++ ewm-service/src/main/java/ru/practicum/dto/event/UpdateEventAdminRequest.java
@@ -1,23 +1,23 @@
-package ru.practicum.api.event;
+package ru.practicum.dto.event;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import lombok.Data;
 import ru.practicum.model.Event;
 import ru.practicum.model.StateAction;
-import ru.practicum.utility.validation.StartTimeValidation;
 
+import javax.validation.constraints.Future;
 import javax.validation.constraints.Size;
 import java.time.LocalDateTime;
 
 @Data
-public class UpdateEventUserRequest {
+public class UpdateEventAdminRequest {
     @Size(min = 20, max = 2000, message = "Поле annotation должно содержать от 20 до 2000 символов")
     private String annotation;
     private Long category;
     @Size(min = 20, max = 7000, message = "Поле description должно содержать от 20 до 2000 символов")
     private String description;
+    @Future
     @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd HH:mm:ss")
-    @StartTimeValidation
     private LocalDateTime eventDate;
     private Event.Location location;
     private Boolean paid;
