+++ ewm-service/src/main/java/ru/practicum/dto/request/ParticipationRequestDto.java
@@ -1,4 +1,4 @@
-package ru.practicum.api.request;
+package ru.practicum.dto.request;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import lombok.Data;
