+++ ewm-service/src/main/java/ru/practicum/service/category/CategoryServiceImpl.java
@@ -5,8 +5,8 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Sort;
 import org.springframework.lang.Nullable;
 import org.springframework.stereotype.Service;
-import ru.practicum.api.category.CategoryDto;
-import ru.practicum.api.category.NewCategoryDto;
+import ru.practicum.dto.category.CategoryDto;
+import ru.practicum.dto.category.NewCategoryDto;
 import ru.practicum.exception.DeleteEntityException;
 import ru.practicum.exception.NotFoundException;
 import ru.practicum.model.Category;
