+++ ewm-service/src/main/java/ru/practicum/service/request/RequestService.java
@@ -1,8 +1,8 @@
 package ru.practicum.service.request;
 
-import ru.practicum.api.request.EventRequestStatusUpdateRequest;
-import ru.practicum.api.request.EventRequestStatusUpdateResult;
-import ru.practicum.api.request.ParticipationRequestDto;
+import ru.practicum.dto.request.EventRequestStatusUpdateRequest;
+import ru.practicum.dto.request.EventRequestStatusUpdateResult;
+import ru.practicum.dto.request.ParticipationRequestDto;
 
 import java.util.List;
