+++ ewm-service/src/main/java/ru/practicum/service/request/RequestServiceImpl.java
@@ -3,9 +3,9 @@ package ru.practicum.service.request;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
-import ru.practicum.api.request.EventRequestStatusUpdateRequest;
-import ru.practicum.api.request.EventRequestStatusUpdateResult;
-import ru.practicum.api.request.ParticipationRequestDto;
+import ru.practicum.dto.request.EventRequestStatusUpdateRequest;
+import ru.practicum.dto.request.EventRequestStatusUpdateResult;
+import ru.practicum.dto.request.ParticipationRequestDto;
 import ru.practicum.exception.ConfirmationException;
 import ru.practicum.exception.ForbiddenActionException;
 import ru.practicum.exception.NotFoundException;
