+++ ewm-service/src/main/java/ru/practicum/service/user/UserServiceImpl.java
@@ -4,8 +4,8 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
-import ru.practicum.api.user.NewUserRequest;
-import ru.practicum.api.user.UserDto;
+import ru.practicum.dto.user.NewUserRequest;
+import ru.practicum.dto.user.UserDto;
 import ru.practicum.exception.NotFoundException;
 import ru.practicum.model.User;
 import ru.practicum.repository.UserRepository;
