+++ ewm-service/src/main/java/ru/practicum/utility/mapper/CompilationMapper.java
@@ -1,7 +1,7 @@
 package ru.practicum.utility.mapper;
 
 import org.mapstruct.Mapper;
-import ru.practicum.api.compilation.CompilationDto;
+import ru.practicum.dto.compilation.CompilationDto;
 import ru.practicum.model.Compilation;
 
 @Mapper(componentModel = "spring")
