+++ ewm-service/src/main/java/ru/practicum/utility/mapper/EventMapper.java
@@ -1,7 +1,7 @@
 package ru.practicum.utility.mapper;
 
 import org.mapstruct.Mapper;
-import ru.practicum.api.event.EventFullDto;
+import ru.practicum.dto.event.EventFullDto;
 import ru.practicum.model.Event;
 
 @Mapper(componentModel = "spring")
