+++ ewm-service/src/main/java/ru/practicum/utility/mapper/UserMapper.java
@@ -1,7 +1,7 @@
 package ru.practicum.utility.mapper;
 
 import org.mapstruct.Mapper;
-import ru.practicum.api.user.UserDto;
+import ru.practicum.dto.user.UserDto;
 import ru.practicum.model.User;
 
 @Mapper(componentModel = "spring")
